alignment_tax_analysis/requirements.txt
# 核心依赖
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0

# 数据处理
tqdm>=4.65.0

# 日志和配置
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0